# backend/internal/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Palworld Scheduler API",
    "version": "1.0.0",
    "description": "Palworld 伺服器維運 API：排程覆寫（開/關/恢復自動）與完整 RCON。\n\n驗證方式：點右上 **Authorize**，在 `X-Auth-Token` 填入你的 token（對應 config.json 的 api.token）。之後即可對各端點按 **Try it out** 直接調用。\n\n備註：訊息類參數（broadcast/shutdown message）內的空白會被 Palworld 截斷，請用底線 `_` 代替空白。"
  },
  "servers": [
    { "url": "/", "description": "目前主機" }
  ],
...
  "components": {
    "securitySchemes": {
      "TokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Auth-Token",
        "description": "填入 config.json 的 api.token"
      }
    },
    "schemas": {
      "SaveSource": {
        "type": "object",
        "properties": {
          "root": { "type": "string", "description": "目前的存檔根目錄（絕對路徑）" },
          "exists": { "type": "boolean", "description": "這個資料夾現在是否存在" },
          "current": { "type": "string", "description": "正在讀的 Level.sav（相對 root）" },
          "mtime": { "type": "integer", "description": "該存檔的最後修改時間（Unix 秒）" },
          "worlds": {
            "type": "array",
            "description": "root 底下找得到的世界",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "string", "description": "世界資料夾名（那串英數）" },
                "path": { "type": "string", "description": "Level.sav 的相對路徑" },
                "size": { "type": "integer", "description": "Level.sav 大小（bytes）" },
                "mtime": { "type": "integer", "description": "最後修改時間（Unix 秒）" }
              }
            }
          }
        }
      },
      "Status": {
        "type": "object",
        "properties": {
          "mode": { "type": "string", "enum": ["AUTO", "FORCE_ON", "FORCE_OFF"], "description": "目前控制模式" },
          "serverRunning": { "type": "boolean", "description": "Palworld 容器是否執行中" },
          "inOpenWindow": { "type": "boolean", "description": "現在是否在開放時段內" },
          "currentLabel": { "type": "string" },
          "currentCloseAt": { "type": "string" },
          "nextOpenLabel": { "type": "string" },
          "nextOpenAt": { "type": "string" },
          "serverTimeLocal": { "type": "string" }
        }
      },
      "ActionResult": {
        "type": "object",
        "properties": {
          "ok": { "type": "boolean" },
          "action": { "type": "string" }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "ok": { "type": "boolean", "example": false },
          "error": { "type": "string" }
        }
      },
...
      "PalsResult": {
        "type": "object",
        "properties": {
          "total_pals": { "type": "integer" },
          "orphan_pals": { "type": "integer" },
          "source": { "type": "string", "description": "解析的存檔相對路徑" },
          "players": { "type": "array", "items": { "$ref": "#/components/schemas/PalPlayer" } }
        }
      }
    }
  },
  "security": [{ "TokenAuth": [] }],
  "paths": {
    "/api/pals/analyze": {
      "post": {
        "tags": ["palsave"],
        "summary": "解析上傳的存檔（不落地）",
        "description": "把一份 Level.sav 的位元組直接送上來解析，回傳與 /api/pals 相同的結構。\n\n整份存檔只會在記憶體裡走一遍（Go 這層串流轉發、palsave 那層從 bytes 解析），伺服器磁碟上不會留下任何副本，也不會動到這台伺服器自己的存檔或快取。\n\n給單機／自架玩家用：他們沒辦法把面板架起來，但手上就有存檔。上限 512 MB；公開入口（nginx）另外對這條路徑限流 3 次/分。",
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": { "schema": { "type": "string", "format": "binary" } },
            "multipart/form-data": { "schema": { "type": "object", "properties": { "file": { "type": "string", "format": "binary" } } } }
          }
        },
        "responses": {
          "200": { "description": "解析結果（多了 source=\"(上傳的存檔)\" 與 uploaded=true）", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PalsResult" } } } },
          "400": { "description": "不是有效的 Palworld 存檔，或沒有收到內容" },
          "413": { "description": "檔案超過 512 MB" },
          "429": { "description": "上傳太頻繁（由 nginx 限流）" },
          "503": { "description": "存檔解析未啟用（config.json 的 palsave.enabled=false）" }
        }
      }
    },
    "/api/palsave/source": {
      "get": {
        "tags": ["palsave"],
        "summary": "查詢目前的存檔來源",
        "description": "回報現在讀的是哪個資料夾、選中哪個世界，以及這個根目錄底下有哪些世界可選。",
        "responses": {
          "200": { "description": "來源資訊", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SaveSource" } } } },
          "502": { "description": "呼叫 palsave 失敗" },
          "503": { "description": "存檔解析未啟用" }
        }
      },
      "post": {
        "tags": ["palsave"],
        "summary": "改用指定的絕對路徑當存檔根目錄",
        "description": "不必再依賴 backend/palworld-data，可以直接指到伺服器上任一存檔資料夾，之後的定時監控與面板資料都改讀那裡。\n\n⚠️ 這是管理者操作，公開入口（nginx）刻意擋掉這條路徑，只能直接打排程器的 9000 埠並帶 API token。\n\n路徑是在「跑排程器的那台機器」上解析的 —— Docker 版要注意那是容器內的路徑，主機資料夾必須先掛進容器。只接受底下含 Pal/Saved/SaveGames 的資料夾；切換成功會立刻作廢預熱快取並重抓。要永久生效請一併寫進 config.json 的 palsave.saveRoot。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["root"],
                "properties": { "root": { "type": "string", "description": "存檔根目錄的絕對路徑", "example": "/palworld" } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "已切換", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SaveSource" } } } },
          "400": { "description": "找不到資料夾，或底下沒有 Pal/Saved/SaveGames" },
          "502": { "description": "呼叫 palsave 失敗" },
          "503": { "description": "存檔解析未啟用" }
        }
      }
    },
...
    "/api/status": {
      "get": {
        "tags": ["schedule"],
        "summary": "查詢目前狀態",
        "responses": {
          "200": { "description": "狀態快照", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Status" } } } },
          "401": { "description": "未授權", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/api/stop": {
      "post": {
        "tags": ["schedule"],
        "summary": "強制關閉（FORCE_OFF）",
        "description": "執行 onManualStop 流程（倒數→存檔→關閉）。忽略排程保持關閉，直到 /api/auto 或 /api/start。",
        "parameters": [
          { "name": "countdown", "in": "query", "required": false, "schema": { "type": "integer", "default": 15 }, "description": "倒數秒數，覆寫流程中的 countdown 步驟" }
        ],
        "responses": {
          "200": { "description": "已切換", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ActionResult" } } } }
        }
      }
    },
...
    "/api/rcon": {
      "post": {
        "tags": ["rcon"],
        "summary": "送出任意 RCON 指令",
        "parameters": [
          { "name": "command", "in": "query", "required": false, "schema": { "type": "string" }, "description": "RCON 指令；也可改用 JSON body" }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "command": { "type": "string", "example": "Broadcast GLHF" } } }
            }
          }
        },
        "responses": {
          "200": { "description": "RCON 回應", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RconResult" } } } },
          "400": { "description": "缺少 command", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "502": { "description": "RCON 失敗", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/broadcast": {
      "post": {
        "tags": ["rcon"],
        "summary": "全服廣播 (Broadcast)",
        "parameters": [
          { "name": "msg", "in": "query", "required": true, "schema": { "type": "string", "example": "Hello_everyone" }, "description": "訊息（空白用底線）" }
        ],
        "responses": {
          "200": { "description": "已送出", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RconResult" } } } },
          "400": { "description": "缺少 msg", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/api/players": {
      "get": {
        "tags": ["rcon"],
        "summary": "在線玩家（改走 REST）",
        "description": "已改用官方 REST 取得結構化玩家資料：RCON 的 ShowPlayers 遇含中文（多位元組）玩家名會因帕魯回應封包長度算錯而讀取逾時，故此端點優先走 REST，回傳與 /api/rest/players 相同。REST 未啟用時才退回 RCON（含中文玩家可能逾時）。",
        "responses": {
          "200": { "description": "玩家列表", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PlayerList" } } } },
          "502": { "description": "REST 呼叫失敗", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/api/kick": {
      "post": {
        "tags": ["rcon"],
        "summary": "踢出玩家 (KickPlayer)",
        "parameters": [
          { "name": "steamid", "in": "query", "required": true, "schema": { "type": "string" }, "description": "SteamID（可先呼叫 /api/players 取得）" }
        ],
        "responses": {
          "200": { "description": "已送出", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RconResult" } } } },
          "400": { "description": "缺少 steamid", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/ban": {
      "post": {
        "tags": ["rcon"],
        "summary": "封鎖玩家 (BanPlayer)",
        "parameters": [
          { "name": "steamid", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "已送出", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RconResult" } } } },
          "400": { "description": "缺少 steamid", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/unban": {
      "post": {
        "tags": ["rcon"],
        "summary": "解除封鎖 (UnBanPlayer)",
        "parameters": [
          { "name": "steamid", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "已送出", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RconResult" } } } },
          "400": { "description": "缺少 steamid", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/shutdown": {
      "post": {
        "tags": ["rcon"],
        "summary": "伺服器倒數關閉 (Shutdown)",
        "description": "Palworld 內建：伺服器自行倒數、存檔、關閉。",
        "parameters": [
          { "name": "seconds", "in": "query", "required": false, "schema": { "type": "integer", "default": 60 } },
          { "name": "message", "in": "query", "required": false, "schema": { "type": "string", "example": "Server_shutting_down" }, "description": "訊息（空白用底線）" }
        ],
        "responses": {
          "200": { "description": "已送出", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RconResult" } } } }
        }
      }
    },
...
    "/api/teleport": {
      "post": {
        "tags": ["rcon"],
        "summary": "傳送 (TeleportToMe / TeleportToPlayer)",
        "description": "需要遊戲內管理員情境；透過 RCON 多半無效，僅作為通道保留。",
        "parameters": [
          { "name": "steamid", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "mode", "in": "query", "required": false, "schema": { "type": "string", "enum": ["tome", "toplayer"], "default": "tome" } }
        ],
        "responses": {
          "200": { "description": "已送出", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RconResult" } } } },
          "400": { "description": "缺少 steamid", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/api/adminpassword": {
      "post": {
        "tags": ["rcon"],
        "summary": "取得管理權限 (AdminPassword)",
        "description": "RCON 連線已完成驗證，一般不需要；僅作為通道保留。",
        "parameters": [
          { "name": "password", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "已送出", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RconResult" } } } },
          "400": { "description": "缺少 password", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/rest/info": {
      "get": {
        "tags": ["rest-api"],
        "summary": "伺服器資訊（REST）",
        "description": "官方 REST GET /v1/api/info，回應為結構化 JSON。",
        "responses": {
          "200": { "description": "伺服器資訊", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ServerInfo" } } } },
          "502": { "description": "REST 呼叫失敗", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "503": { "description": "REST 整合未啟用", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/rest/players": {
      "get": {
        "tags": ["rest-api"],
        "summary": "在線玩家（REST，結構化）",
        "description": "官方 REST GET /v1/api/players，含等級、座標、ping、建築數等；userId 可用於 kick/ban。",
        "responses": {
          "200": { "description": "玩家列表", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PlayerList" } } } },
          "502": { "description": "REST 呼叫失敗", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "503": { "description": "REST 整合未啟用", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/api/rest/settings": {
      "get": {
        "tags": ["rest-api"],
        "summary": "伺服器設定（REST）",
        "description": "官方 REST GET /v1/api/settings，回傳目前生效的 OptionSettings。",
        "responses": {
          "200": { "description": "設定物件", "content": { "application/json": { "schema": { "type": "object" } } } },
          "502": { "description": "REST 呼叫失敗", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "503": { "description": "REST 整合未啟用", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/rest/announce": {
      "post": {
        "tags": ["rest-api"],
        "summary": "廣播訊息（REST）",
        "description": "官方 REST POST /v1/api/announce。可用 ?msg= 或 JSON body。REST 版本訊息可含空白，不必用底線。",
        "parameters": [
          { "name": "msg", "in": "query", "required": false, "schema": { "type": "string", "example": "Hello everyone" } }
        ],
        "requestBody": {
          "required": false,
          "content": { "application/json": { "schema": { "type": "object", "properties": { "message": { "type": "string", "example": "Hello everyone" } } } } }
        },
        "responses": {
          "200": { "description": "已廣播", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RestAction" } } } },
          "400": { "description": "缺少訊息", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "502": { "description": "REST 呼叫失敗", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "503": { "description": "REST 整合未啟用", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/rest/kick": {
      "post": {
        "tags": ["rest-api"],
        "summary": "踢出玩家（REST）",
        "description": "官方 REST POST /v1/api/kick。",
        "parameters": [
          { "name": "userid", "in": "query", "required": true, "schema": { "type": "string" }, "description": "玩家 userId（由 /api/rest/players 取得，如 steam_0000...）" },
          { "name": "message", "in": "query", "required": false, "schema": { "type": "string" }, "description": "顯示原因" }
        ],
        "responses": {
          "200": { "description": "已踢出", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RestAction" } } } },
          "400": { "description": "缺少 userid", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "502": { "description": "REST 呼叫失敗", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "503": { "description": "REST 整合未啟用", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/rest/ban": {
      "post": {
        "tags": ["rest-api"],
        "summary": "封鎖玩家（REST）",
        "description": "官方 REST POST /v1/api/ban。",
        "parameters": [
          { "name": "userid", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "message", "in": "query", "required": false, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "已封鎖", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RestAction" } } } },
          "400": { "description": "缺少 userid", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "502": { "description": "REST 呼叫失敗", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "503": { "description": "REST 整合未啟用", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/rest/unban": {
      "post": {
        "tags": ["rest-api"],
        "summary": "解除封鎖（REST）",
        "description": "官方 REST POST /v1/api/unban。",
        "parameters": [
          { "name": "userid", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "已解封", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RestAction" } } } },
          "400": { "description": "缺少 userid", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "502": { "description": "REST 呼叫失敗", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "503": { "description": "REST 整合未啟用", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/rest/save": {
      "post": {
        "tags": ["rest-api"],
        "summary": "存檔（REST）",
        "description": "官方 REST POST /v1/api/save。",
        "responses": {
          "200": { "description": "已存檔", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RestAction" } } } },
          "502": { "description": "REST 呼叫失敗", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "503": { "description": "REST 整合未啟用", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/rest/shutdown": {
      "post": {
        "tags": ["rest-api"],
        "summary": "倒數關閉（REST）",
        "description": "官方 REST POST /v1/api/shutdown：伺服器倒數 seconds 秒後關閉並顯示 message。",
        "parameters": [
          { "name": "seconds", "in": "query", "required": false, "schema": { "type": "integer", "default": 60 } },
          { "name": "message", "in": "query", "required": false, "schema": { "type": "string", "example": "Server shutting down" } }
        ],
        "responses": {
          "200": { "description": "已排定關閉", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RestAction" } } } },
          "502": { "description": "REST 呼叫失敗", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "503": { "description": "REST 整合未啟用", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/rest/stop": {
      "post": {
        "tags": ["rest-api"],
        "summary": "立即強制停止（REST）",
        "description": "官方 REST POST /v1/api/stop：不倒數、不通知，立即停止伺服器。",
        "responses": {
          "200": { "description": "已停止", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RestAction" } } } },
          "502": { "description": "REST 呼叫失敗", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "503": { "description": "REST 整合未啟用", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/pals/players": {
      "get": {
        "tags": ["palsave"],
        "summary": "玩家清單摘要（先取得 UID）",
        "description": "只回玩家層級資料（name、uid、level、exp、hp、pal_count…），不含每隻帕魯，回應輕量。先用這個取得玩家 UID，再用 /api/pals?q=<UID> 查完整帕魯。",
        "parameters": [
          { "name": "uuid", "in": "query", "required": false, "schema": { "type": "string" }, "description": "以存檔內玩家 UUID（或名稱）部分比對過濾" }
        ],
        "responses": {
          "200": { "description": "玩家清單（players 內每筆為 PalPlayer 去掉 pals 欄位）", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PalsResult" } } } },
          "502": { "description": "呼叫 palsave 失敗", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "503": { "description": "存檔解析未啟用", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/pals": {
      "get": {
        "tags": ["palsave"],
        "summary": "玩家帕魯資料（解析存檔）",
        "description": "經 palsave sidecar 解析目前世界的 Level.sav，回傳每位玩家擁有的帕魯（種類中英文名、等級、性別、個體值、α/Lucky）。RCON/REST 皆無此資料。結果依存檔 mtime 快取。",
        "parameters": [
          { "name": "uuid", "in": "query", "required": false, "schema": { "type": "string" }, "description": "只查符合的玩家：存檔內玩家 UUID（或名稱），部分比對" }
        ],
        "responses": {
          "200": { "description": "玩家帕魯資料", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PalsResult" } } } },
          "502": { "description": "呼叫 palsave 失敗", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "503": { "description": "存檔解析未啟用", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    }
  }
}